feat(main): Fix's to front complete deliry/customer

+++ src/services/productService.ts
@@ -1,5 +1,5 @@
 import { ICreateProductDTO } from "../dtos/ICreateProductDTO";
-import { NotFoundError } from "../helpers/apiErros";
+import { BadRequestError, NotFoundError } from "../helpers/apiErros";
 import { ProductRepository } from "../repositories/ProductRepository";
 
 class ProductService {
@@ -29,12 +29,20 @@ class ProductService {
 
   async update(id: string, data: ICreateProductDTO) {
     const product = await this.productRepository.showById(id);
-
+    
     if (!product) {
       throw new NotFoundError('Product not found');
     }
 
-    const newProduct = await this.productRepository.update(id, data);
+    const productWithout_count = {
+      name: data.name,
+      description: data.description,
+      price: data.price,
+      quantity: data.quantity,
+      sellerId: data.sellerId
+    }
+
+    const newProduct = await this.productRepository.update(id, productWithout_count);
 
     return newProduct;
   }
@@ -46,6 +54,10 @@ class ProductService {
       throw new NotFoundError('Product not found');
     }
 
+    if (product._count.DeliveryRequest > 0) {
+      throw new BadRequestError('Product has delivery requests');
+    }
+    
     await this.productRepository.delete(id);
   }
 
@@ -62,15 +74,21 @@ class ProductService {
       throw new NotFoundError('Product not found');
     }
 
-    console.log(product.quantity, quantity)
-
     if(product.quantity < quantity) {
       throw new NotFoundError('Insufficient quantity');
     }
 
     const newQuantity = product.quantity - quantity;
 
-    const newProduct = await this.productRepository.update(id, { ...product, quantity: newQuantity });
+    const productWithout_count = {
+      name: product.name,
+      description: product.description,
+      price: product.price,
+      quantity: product.quantity,
+      sellerId: product.sellerId
+    }
+
+    const newProduct = await this.productRepository.update(id, { ...productWithout_count, quantity: newQuantity });
 
     return newProduct;
   }
